backend_pipelines/sentiment_analysis/requirements.txt
google-cloud-firestore
vaderSentiment==3.3.2
